Dereferenced possible null references in DeleteFood.cshtml.cs

+++ src/Pages/Product/DeleteFood.cshtml.cs
@@ -38,9 +38,12 @@ namespace ContosoCrafts.WebSite.Pages.Product
         /// <param name="id"></param>
         public void OnGet(string id)
         {
-            var checknull = ProductService.GetFood().FirstOrDefault(m => m.Id.Equals(id));
+            var checknull = ProductService.GetFood().FirstOrDefault(m => m.Id != null && m.Id.Equals(id));
             if (checknull is not null) {
                 Food = checknull;
+            } else {
+                // redirect to error page
+
             }
             Products = ProductService.GetProducts();
         }
@@ -59,7 +62,10 @@ namespace ContosoCrafts.WebSite.Pages.Product
                 return Page();
             }
 
-            ProductService.DeleteFoodData(Food.Id);
+            if(Food.Id is not null)
+            {
+                ProductService.DeleteFoodData(Food.Id);
+            }
 
             return RedirectToPage("./ProductIndex");
         }
